+++ sgdbBanco/src/main/java/com/miguel/seatcode/bootcamp/sgdb/banco/clases/HistoricoActions.java
@@ -1,5 +1,5 @@
 package com.miguel.seatcode.bootcamp.sgdb.banco.clases;
 
 public interface HistoricoActions {
-    public String crearHistorico(String movimiento, float balance, int fk_id_cuenta);
+    public String crearHistorico(String movimiento, float balance, Long fk_id_cuenta);
 }
